chess-backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-chess==1.999